pyproject.toml
[project]
name = "ai-chat"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chardet>=7.4.0.post2",
    "duckduckgo-search>=8.1.1",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "markdown>=3.10.2",
    "openpyxl>=3.1.5",
    "pillow>=12.2.0",
    "pymupdf>=1.27.2.2",
    "pymupdf4llm>=1.27.2.2",
    "python-arango>=8.3.1",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "starlette>=1.0.0",
    "uvicorn>=0.44.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

[dependency-groups]
dev = [
    "ruff>=0.9",
    "pytest>=8.0",
    "bandit>=1.8",
    "pyright>=1.1",
    "pytest-asyncio>=1.3.0",
]
